implemented LabelButtonV2 into the rest of the app. Added Barbecue depencences in Constellio Gradle. Fixed bug in LabelButtonV2

+++ app/src/com/constellio/app/modules/rm/wrappers/PrintableLabel.java
@@ -31,11 +31,11 @@ public class PrintableLabel extends Printable implements Dimensionnable {
     }
 
     public int getColonne() {
-        return get(COLONNE);
+        return  ((Double) get(COLONNE)).intValue();
     }
 
     public int getLigne() {
-        return get(LIGNE);
+        return ((Double) get(LIGNE)).intValue();
     }
 
     @Override
